backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "pipecat-ai[gemini,daily,google]>=0.0.74",
    "google-api-core>=2.0.0",
    "google-auth>=2.0.0",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
]
